SpikingRasters

Documentation for SpikingRasters.

SpikingRasters.count_spikes

SpikingRasters.detect_spikes

SpikingRasters.generate_spike_raster

SpikingRasters.count_spikes — Method

count_spikes(Y::Vector{R},dt::R,Tend::R;Tstart::R=0.0) where R

Arguments

Y::Vector{<:Real} : vector of spike times

dt::Real : time bin size

Tend::Real : end time for the raster

Optional argument

Tstart::Real=0.0 : start time for the raster

Returns

binned_spikes::Vector{<:Integer} : binned_spikes[k] is the number of spikes that occur in the timebin k (i.e. between Tstart + (k-1)*dt and Tstart + k*dt)

SpikingRasters.detect_spikes — Method

detect_spikes(Y::Vector{R},dt::R,Tend::R;Tstart::R=0.0) where R

Arguments

Y::Vector{<:Real} : vector of spike times

dt::Real : time bin size

Tend::Real : end time for the raster

Optional argument

Tstart::Real=0.0 : start time for the raster

Returns

detected_spikes::Vector{<:Integer} : detected_spike[k] is true if at least one spike has been found.

SpikingRasters.generate_spike_raster — Method

generatespikeraster(trains::SpikeTrainCollection, dt::Real; spikesize::Integer = 5, spikeseparator::Integer = 1, backgroundcolor::Color=RGB(1.,1.,1.), spikecolors::Union{C,Vector{C}}=RGB(0.,0.0,0.0), max_size::Real=1E4) where C<:Color

Draws a matrix that contains the raster plot of the spike train.

Arguments

Trains : The spike trains.

dt : time interval representing one pixel horizontally

Optional arguments

spike_size::Integer : heigh of spike (in pixels)

spike_separator::Integer : space between spikes, and vertical padding

background_color::Color : self-explanatory

spike_colors::Union{Color,Vector{Color}} : if a single color, color of all spikes, if vector of colors, color for each neuron (length should be same as number of neurons)

max_size::Integer : throws an error if image is larger than this number (in pixels)

Returns

raster_matrix::Matrix{Color} you can save it as a png file